trying to build from ticktack-builder (electron entry needs to be index.js?)
